rename button file
